+++ lib/git.js
@@ -41,7 +41,7 @@ function init() {
             console.log(`There was an error initializing the repository: ${error}`);
         }
         finally {
-            return Promise.resolve("Initialization step complete...");
+            console.log("Initialization step complete...");
         }
     });
 }
@@ -77,7 +77,7 @@ function generateBranch() {
             core_1.setFailed(`There was an error creating the deployment branch: ${error} ❌`);
         }
         finally {
-            return Promise.resolve("Deployment branch creation step complete... ✅");
+            console.log("Deployment branch creation step complete... ✅");
         }
     });
 }
@@ -125,7 +125,7 @@ function deploy() {
         const hasFilesToCommit = yield execute_1.execute(`git status --porcelain`, temporaryDeploymentDirectory);
         if (!hasFilesToCommit && !constants_1.action.isTest) {
             console.log("There is nothing to commit. Exiting... ✅");
-            return Promise.resolve();
+            return Promise.resolve("Exiting early...");
         }
         // Commits to GitHub.
         yield execute_1.execute(`git add --all .`, temporaryDeploymentDirectory);
